Update the list of constants
TRUE and FALSE are in the constant.language.lsl category alone. They are treated by the compiler slightly differently (they can't have a minus sign if used in the globals section, while the others can). The rest are in the support.constant.lsl category. This distinction may be artificial and unnecessary.

This is the list of constants as of 2017-09-06.