Update the list of constants, but some are still missing.
TRUE and FALSE are in the constant.language.lsl category alone. They are treated by the compiler slightly differently (they can't have a minus sign if used in the globals section, while the others can). The rest are in the support.constant.lsl category. This distinction may be artificial and unnecessary.

This is the list of constants as of 2017-09-06 with these exceptions:

- Deprecated constants (because they have a misleading name, or for other reasons)
  PSYS_SRC_INNERANGLE
  PSYS_SRC_OUTERANGLE
  VEHICLE_FLAG_NO_FLY_UP
  ATTACH_RPEC
  ATTACH_LPEC
  LAND_SMALL_BRUSH
  LAND_MEDIUM_BRUSH
  LAND_LARGE_BRUSH
  DATA_RATING
  PRIM_CAST_SHADOWS
  PRIM_MATERIAL_LIGHT
- Unimplemented constants (they exist but they serve no purpose, even if they imply one):
  PERMISSION_REMAP_CONTROLS
  PERMISSION_RELEASE_OWNERSHIP
  PERMISSION_CHANGE_JOINTS
  PERMISSION_CHANGE_PERMISSIONS
  PSYS_SRC_OBJ_REL_MASK